+++ fanboi2/tests/test_views.py
@@ -750,6 +750,24 @@ class TestBoardViews(ViewMixin, unittest.TestCase):
         result_.assert_called_with('dummy')
         self.assertEqual(response.status, '422 Unprocessable Entity')
 
+    @unittest.mock.patch('fanboi2.tasks.celery.AsyncResult')
+    def test_board_new_get_proxy_rejected(self, result_):
+        from fanboi2.views.boards import board_new_get
+        board = self._makeBoard(title='Foobar', slug='foobar')
+        result_.return_value = DummyAsyncResult('dummy', 'success', [
+            'failure',
+            'proxy_rejected'])
+
+        request = self._GET()
+        request.matchdict['board'] = board.slug
+        request.params['task'] = 'dummy'
+        config = self._makeConfig(request, self._makeRegistry())
+        config.testing_add_renderer('boards/error_proxy.mako')
+
+        response = board_new_get(request)
+        result_.assert_called_with('dummy')
+        self.assertEqual(response.status, '422 Unprocessable Entity')
+
     def test_board_new_get_not_found(self):
         from sqlalchemy.orm.exc import NoResultFound
         from fanboi2.views.boards import board_new_get
@@ -973,6 +991,28 @@ class TestBoardViews(ViewMixin, unittest.TestCase):
         request.params['task'] = 'dummy'
         config = self._makeConfig(request, self._makeRegistry())
         config.testing_add_renderer('topics/error_status.mako')
+
+        response = topic_show_get(request)
+        result_.assert_called_with('dummy')
+        self.assertEqual(response.status, '422 Unprocessable Entity')
+
+    @unittest.mock.patch('fanboi2.tasks.celery.AsyncResult')
+    def test_topic_show_get_proxy_rejected(self, result_):
+        from fanboi2.views.boards import topic_show_get
+        board = self._makeBoard(title='Foobar', slug='foobar')
+        topic = self._makeTopic(board=board, title='Foobar')
+        self._makePost(topic=topic, body='Lorem ipsum')
+        self._makePost(topic=topic, body='Dolor sit amet')
+        result_.return_value = DummyAsyncResult('dummy', 'success', [
+            'failure',
+            'proxy_rejected'])
+
+        request = self._GET()
+        request.matchdict['board'] = board.slug
+        request.matchdict['topic'] = topic.id
+        request.params['task'] = 'dummy'
+        config = self._makeConfig(request, self._makeRegistry())
+        config.testing_add_renderer('topics/error_proxy.mako')
 
         response = topic_show_get(request)
         result_.assert_called_with('dummy')